Different view based on whether the CSV is loaded
