+++ blackheap-frontend/src/routes/Evaluation/EvaluationDashboard.tsx
@@ -1,0 +1,12 @@
+import PreloadeeRecords from "../../types/PreloadeeRecords";
+
+// TODO any
+const EvaluationDashboard = ({preloadeeRecords}: {preloadeeRecords: PreloadeeRecords}) => {
+    return (
+        <>
+        {preloadeeRecords.map(r => r.io_type).join(",")}
+        </>
+    );
+}
+
+export default EvaluationDashboard;
